consolidated files, made each able to talk to one another, tested a few times to make sure communication is working.
